+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # wrapper_tags change log
 
 
-### 1.2.2 (2017-12-?)
+### 1.2.2 (2017-12-9)
 
 * fix: copy and move mode do not show validation, colors and indents
 * fix: responsiveness of opening wrapper-tags element edit panel
@@ -9,6 +9,7 @@
 * style attribute can be set on every html tag
 * new setting: hide validation status
 * full css-es in debug mode 
+* composer.json
 
 ### 1.2.0 (2017-12-7)